Fix confirm finansial and pendukung

+++ listdata.php
@@ -89,7 +89,7 @@ function listNonNeraca() {
 	$datagrid = new simbio_datagrid();
 	$table_spec = 'non_coa as c LEFT JOIN koperasi as k ON c.idkoperasi = k.idkoperasi';
 	$datagrid->setSQLColumn('CONCAT(\'<a href="datacenter-entrydata-pendukung.php?nid=\',c.idnon_coa,\'">Edit</a>\') as \'&nbsp;\'',
-		'CONCAT(\'<a href="datacenter-delete.php?oid=\',c.idnon_coa,\'">Hapus</a>\') as \'&nbsp;\'',
+		'CONCAT(\'<a href="datacenter-delete.php?oid=\',c.idnon_coa,\'" class="confirm">Hapus</a>\') as \'&nbsp;\'',
 		'k.nama AS \'Koperasi\'', 'c.periode AS \'Periode Laporan\'');
 	$datagrid->table_header_attr = 'style="font-weight: bold; color:rgb(255,255,255); background-color:cyan; vertical-align:middle;"';
 	$datagrid->debug = true;
@@ -108,7 +108,7 @@ function listFinasial() {
 	$table_spec = 'harian as h LEFT JOIN koperasi as k ON h.idkoperasi = k.idkoperasi LEFT JOIN tipe_koperasi as t ON k.jenis = t.idtipe_koperasi';
 	$datagrid->table_header_attr = 'style="font-weight: bold; color:rgb(255,255,255); background-color:cyan; vertical-align:middle;"';
     $datagrid->setSQLColumn('CONCAT(\'<a href="datacenter-entrydata-finansial.php?nid=\',h.idday,\'">Edit</a>\') as \'&nbsp;\'',
-		'CONCAT(\'<a href="datacenter-delete.php?fid=\',h.idday,\'">Hapus</a>\') as \'&nbsp;\'',
+		'CONCAT(\'<a href="datacenter-delete.php?fid=\',h.idday,\'" class="confirm">Hapus</a>\') as \'&nbsp;\'',
         'k.nama AS \'Koperasi\'', 'k.sandilembaga AS \'Sandi Lembaga\'',
         't.jenis AS \'Tipe Koperasi\'', 'h.periode AS \'Periode Lap\'');
     if ($group ==2 ) {
